requirements.txt
flask==2.3.3
africastalking==0.1.6
psycopg2-binary==2.9.9
scikit-learn==1.3.0
pandas==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
numpy==1.24.3
sqlite3==2.6.0 